gpio: Add a singleton-pattern like CreatePin for class
Added a static PinCreate function for which shall be used to
create an instance of a pin. The constructor is private now.

PinCreate checks if an instance of the pin already exists and then either
uses the constructor to create an instance or return the already existing
one if it exists.