major restructure: Do major restructure to define a base class for pin.
Create a PinBase class which acts as a base class for any type of
pin that the user wants to create. GPIO, Exti, Usart etc will
all be a derived class of PinBase. THis has been done to enable
extensability. GPIO and Exti have been seperated as a result of
this enablement.

WARNING: This needs to be tested since this can be a major
source of bugs for functionalities that have been developed before
this commit.