DriveFactory: Switch to smart pointers
There was an elusive bug. A pointer to pwm_timer was being created in local
scope and it went out of scope when the flow returned back to caller, making
the system behave in an unexpeced way. So fixed it by dynamically allocating
pwm_timer.

But ultimately decided to use smart pointer to avoid such problems in the future.